refactor(common-utils): Move recursive deletion tests to `funTest`
Prepare for adding a more heavy-weight test.
