+++ utils/common/src/test/kotlin/ExtensionsTest.kt
@@ -99,37 +99,6 @@ class ExtensionsTest : WordSpec({
         }
     }
 
-    "File.safeDeleteRecursively()" should {
-        "be able to delete files that are not writable" {
-            val dir = tempdir().apply {
-                resolve("read-only.txt").apply {
-                    writeText("Hello!")
-                    check(setWritable(false))
-                }
-            }
-
-            shouldNotThrow<IOException> {
-                dir.safeDeleteRecursively(force = true)
-            }
-
-            dir.exists() shouldBe false
-        }
-
-        "delete empty parent directories below a base directory" {
-            val tempDir = tempdir()
-            val baseDir = tempDir.resolve("a/basedir")
-            val deleteDir = baseDir.resolve("c/delete").apply { safeMkdirs() }
-
-            shouldNotThrow<IOException> {
-                deleteDir.safeDeleteRecursively(force = true, baseDir)
-            }
-
-            deleteDir.exists() shouldBe false
-            deleteDir.parentFile.exists() shouldBe false
-            baseDir.exists() shouldBe true
-        }
-    }
-
     "File.expandTilde()" should {
         "expand the path if the SHELL environment variable is set".config(enabled = Os.env["SHELL"] != null) {
             File("~/Desktop").expandTilde() shouldBe Os.userHomeDirectory.resolve("Desktop")
